take most recognized digit

+++ DigitIsolation.py
@@ -42,6 +42,7 @@ class DigitIsolation:
 
     @staticmethod
     def four_point_transform(image, pts):
+
         # obtain a consistent order of the points and unpack them
         # individually
         tr, tl, bl, br = DigitIsolation.order_points(pts)
@@ -74,7 +75,7 @@ class DigitIsolation:
             [0, 0],
             [maxWidth - 1, 0],
             [maxWidth - 1, maxHeight - 1],
-            [0, maxHeight - 1]], dtype="int8")
+            [0, maxHeight - 1]], dtype="float32")
 
         # compute the perspective transform matrix and then apply it
         M = cv2.getPerspectiveTransform(rect, dst)
